refactor: network 관심사별 분리
